Factor out dates.py.
